Feat/multitouch

+++ aui.views/src/AUI/Event/APointerReleasedEvent.h
@@ -25,7 +25,7 @@ struct APointerReleasedEvent {
     /**
      * @brief Where does the event occurred.
      */
-    glm::ivec2 position;
+    glm::vec2 position;
 
     /**
      * @brief Which button of the pointing device is triggered the event (AInput::LBUTTON if not present) or finger.
